# video/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /douyin/comment/action:
        post:
            tags:
                - CommentService
            description: 发布/删除评论
            operationId: CommentService_Comment
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/comment.api.comment.v1.Douyin_comment_send_request'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/comment.api.comment.v1.Douyin_comment_send_response'
    /douyin/comment/count/video:
        get:
            tags:
                - CommentService
            description: 获取视频评论总数
            operationId: CommentService_GetCommentCntByVId
            parameters:
                - name: videoId
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/comment.api.comment.v1.GetCommentCntByVIdResp'
    /douyin/comment/list:
        get:
            tags:
                - CommentService
            description: 获取评论列表
            operationId: CommentService_CommentList
            parameters:
                - name: token
                  in: query
                  schema:
                    type: string
                - name: videoId
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/comment.api.comment.v1.Douyin_comment_list_response'
...
    /douyin/relation/action:
        post:
            tags:
                - RelationService
            description: 关注/取消关注操作
            operationId: RelationService_Relation
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/relation.api.relation.v1.Douyin_relation_action_request'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/relation.api.relation.v1.Douyin_relation_action_response'
    /douyin/relation/follow/list:
        get:
            tags:
                - RelationService
            description: 获取关注列表
            operationId: RelationService_RelationFollowList
            parameters:
                - name: userId
                  in: query
                  schema:
                    type: string
                - name: token
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/relation.api.relation.v1.Douyin_relation_follow_list_response'
    /douyin/relation/followcnt:
        get:
            tags:
                - RelationService
            description: 获取关注与粉丝数量
            operationId: RelationService_FollowCnt
            parameters:
                - name: userId
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/relation.api.relation.v1.FollowCnt_response'
    /douyin/relation/follower/list:
        get:
            tags:
                - RelationService
            description: 获取粉丝列表
            operationId: RelationService_RelationFollowerList
            parameters:
                - name: userId
                  in: query
                  schema:
                    type: string
                - name: token
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/relation.api.relation.v1.Douyin_relation_follower_list_response'
    /douyin/relation/friend/list:
        get:
            tags:
                - RelationService
            description: 获取好友列表
            operationId: RelationService_FriendList
            parameters:
                - name: userId
                  in: query
                  schema:
                    type: string
                - name: token
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/relation.api.relation.v1.Douyin_relation_friend_list_response'
    /douyin/relation/isfollow:
        get:
            tags:
                - RelationService
            description: 根据 userId 和 authorId 查询用户是否关注作者
            operationId: RelationService_IsFollow
            parameters:
                - name: userId
                  in: query
                  schema:
                    type: string
                - name: authorId
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/relation.api.relation.v1.IsFollow_response'
...
components:
    schemas:
        comment.api.comment.v1.Comment:
            type: object
            properties:
                id:
                    type: string
                user:
                    $ref: '#/components/schemas/comment.api.comment.v1.User'
                content:
                    type: string
                createDate:
                    type: string
        comment.api.comment.v1.Douyin_comment_list_response:
            type: object
            properties:
                statusCode:
                    type: integer
                    format: int32
                statusMsg:
                    type: string
                commentList:
                    type: array
                    items:
                        $ref: '#/components/schemas/comment.api.comment.v1.Comment'
        comment.api.comment.v1.Douyin_comment_send_request:
            type: object
            properties:
                token:
                    type: string
                videoId:
                    type: string
                actionType:
                    type: string
                commentText:
                    type: string
                commentId:
                    type: string
        comment.api.comment.v1.Douyin_comment_send_response:
            type: object
            properties:
                statusCode:
                    type: integer
                    format: int32
                statusMsg:
                    type: string
                comment:
                    $ref: '#/components/schemas/comment.api.comment.v1.Comment'
        comment.api.comment.v1.GetCommentCntByVIdResp:
            type: object
            properties:
                commentCount:
                    type: string
        comment.api.comment.v1.User:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                followCount:
                    type: string
                followerCount:
                    type: string
                isFollow:
                    type: boolean
                avatar:
                    type: string
                backgroundImage:
                    type: string
                signature:
                    type: string
                totalFavorited:
                    type: string
                workCount:
                    type: string
                favoriteCount:
                    type: string
...
        relation.api.relation.v1.Douyin_relation_action_request:
            type: object
            properties:
                token:
                    type: string
                toUserId:
                    type: string
                actionType:
                    type: integer
                    format: int32
        relation.api.relation.v1.Douyin_relation_action_response:
            type: object
            properties:
                statusCode:
                    type: integer
                    format: int32
                statusMsg:
                    type: string
        relation.api.relation.v1.Douyin_relation_follow_list_response:
            type: object
            properties:
                statusCode:
                    type: integer
                    format: int32
                statusMsg:
                    type: string
                userList:
                    type: array
                    items:
                        $ref: '#/components/schemas/relation.api.relation.v1.User'
        relation.api.relation.v1.Douyin_relation_follower_list_response:
            type: object
            properties:
                statusCode:
                    type: integer
                    format: int32
                statusMsg:
                    type: string
                userList:
                    type: array
                    items:
                        $ref: '#/components/schemas/relation.api.relation.v1.User'
        relation.api.relation.v1.Douyin_relation_friend_list_response:
            type: object
            properties:
                statusCode:
                    type: integer
                    format: int32
                statusMsg:
                    type: string
                userList:
                    type: array
                    items:
                        $ref: '#/components/schemas/relation.api.relation.v1.FriendUser'
        relation.api.relation.v1.FollowCnt_response:
            type: object
            properties:
                followCnt:
                    type: string
                followerCnt:
                    type: string
        relation.api.relation.v1.FriendUser:
            type: object
            properties:
                user:
                    $ref: '#/components/schemas/relation.api.relation.v1.User'
                message:
                    type: string
                msgType:
                    type: string
        relation.api.relation.v1.IsFollow_response:
            type: object
            properties:
                isFollow:
                    type: boolean
        relation.api.relation.v1.User:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                followCount:
                    type: string
                followerCount:
                    type: string
                isFollow:
                    type: boolean
                avatar:
                    type: string
                backgroundImage:
                    type: string
                signature:
                    type: string
                totalFavorited:
                    type: string
                workCount:
                    type: string
                favoriteCount:
                    type: string
...